requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
psycopg[binary]>=3.1.0
sqlparse>=0.4.4
sqlglot>=18.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0
pyyaml>=6.0.1
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
aiohttp>=3.8.0
openai>=1.0.0
anthropic>=0.7.0
httpx[socks]>=0.24.0
